feat: pg

+++ apps/api-journeys/src/app/modules/block/image/image.resolver.ts
@@ -4,10 +4,10 @@ import { UserInputError } from 'apollo-server-errors'
 import { encode } from 'blurhash'
 import fetch from 'node-fetch'
 import sharp from 'sharp'
+import { omit } from 'lodash'
 
 import { BlockService } from '../block.service'
 import {
-  CardBlock,
   ImageBlock,
   ImageBlockCreateInput,
   ImageBlockUpdateInput,
@@ -15,6 +15,7 @@ import {
   UserJourneyRole
 } from '../../../__generated__/graphql'
 import { RoleGuard } from '../../../lib/roleGuard/roleGuard'
+import { PrismaService } from '../../../lib/prisma.service'
 
 export async function handleImage(
   input
@@ -70,7 +71,11 @@ export async function handleImage(
 
 @Resolver('ImageBlock')
 export class ImageBlockResolver {
-  constructor(private readonly blockService: BlockService) {}
+  constructor(
+    private readonly blockService: BlockService,
+    private readonly prismaService: PrismaService
+  ) {}
+
   @Mutation()
   @UseGuards(
     RoleGuard('input.journeyId', [
@@ -86,22 +91,32 @@ export class ImageBlockResolver {
 
     if (block.isCover === true) {
       const coverBlock: ImageBlock = await this.blockService.save({
-        ...block,
-        __typename: 'ImageBlock',
+        ...omit(block, ['journeyId', '__typename']),
+        id: block.id ?? undefined,
+        typename: 'ImageBlock',
+        journey: { connect: { id: block.journeyId } },
         parentOrder: null
       })
-      const parentBlock: CardBlock = await this.blockService.get(
-        block.parentBlockId
-      )
+      const parentBlock = await this.prismaService.block.findUnique({
+        where: {
+          id: block.parentBlockId
+        }
+      })
+
+      if (parentBlock == null) {
+        throw new Error('Parent block not found')
+      }
 
       await this.blockService.update(parentBlock.id, {
         coverBlockId: coverBlock.id
       })
       // Delete old coverBlock
       if (parentBlock.coverBlockId != null) {
-        const coverBlockToDelete = await this.blockService.get(
-          parentBlock.coverBlockId
-        )
+        const coverBlockToDelete = await this.prismaService.block.findUnique({
+          where: {
+            id: parentBlock.coverBlockId
+          }
+        })
         if (coverBlockToDelete != null) {
           await this.blockService.removeBlockAndChildren(
             parentBlock.coverBlockId,
@@ -118,8 +133,10 @@ export class ImageBlockResolver {
       block.parentBlockId
     )
     return await this.blockService.save({
-      ...block,
-      __typename: 'ImageBlock',
+      ...omit(block, ['journeyId', '__typename']),
+      id: block.id ?? undefined,
+      typename: 'ImageBlock',
+      journey: { connect: { id: block.journeyId } },
       parentOrder: siblings.length
     })
   }
@@ -138,6 +155,6 @@ export class ImageBlockResolver {
     @Args('input') input: ImageBlockUpdateInput
   ): Promise<ImageBlock> {
     const block = await handleImage(input)
-    return await this.blockService.update(id, block)
+    return await this.blockService.update(id, { ...block, id })
   }
 }
